+++ lib/src/algorithms/argon2/security.dart
@@ -1,16 +1,13 @@
 // Copyright (c) 2023, Sudipto Chandra
 // All rights reserved. Check LICENSE file for details.
 
-import 'dart:math' show min;
-
 import 'argon2.dart';
 
 /// This contains some recommended values of memory, iteration and parallelism
 /// values for [Argon2] algorithm.
 ///
 /// It is best to try out different combinations of these values to achieve the
-/// desired runtime on a target machine. You can use the [tuneArgon2Security]
-/// method for tuning out the best parameters.
+/// desired runtime on a target machine.
 class Argon2Security {
   final String name;
 
@@ -63,108 +60,4 @@ class Argon2Security {
   ///
   /// [rfc]: https://www.ietf.org/rfc/rfc9106.html
   static const strong = Argon2Security('strong', m: 1 << 21, p: 4, t: 1);
-}
-
-/// Find the Argon2 parameters that can be used to encode password in
-/// [desiredRuntime] time on the current device.
-///
-/// This function may take up to `50 * desiredRuntime` time to compute.
-Future<Argon2Security> tuneArgon2Security(
-  Duration desiredRuntime, {
-  int strictness = 10,
-  int saltLength = 16,
-  int hashLength = 32,
-  int maxMemoryAsPowerOf2 = 22,
-  Argon2Type type = Argon2Type.argon2i,
-  Argon2Version version = Argon2Version.v13,
-  bool verbose = false,
-}) async {
-  if (strictness < 1) {
-    throw ArgumentError('Strictness value must be at least 1');
-  }
-  if (maxMemoryAsPowerOf2 < 3) {
-    throw ArgumentError('Max memory as power of 2 must be at least 3');
-  }
-
-  var watch = Stopwatch()..start();
-  var salt = List.filled(saltLength, 1);
-  var password = List.filled(saltLength << 2, 2);
-  int target = desiredRuntime.inMicroseconds;
-
-  // maximize memory
-  int pow = 3, memory, lanes = 1, passes = 1;
-  for (; pow <= maxMemoryAsPowerOf2; pow++) {
-    memory = 1 << pow;
-    lanes = min(16, memory >>> 3);
-    var samples = List.generate(10, (_) {
-      var f = Argon2(
-        salt: salt,
-        hashLength: hashLength,
-        type: type,
-        version: version,
-        iterations: passes,
-        parallelism: lanes,
-        memorySizeKB: memory,
-      );
-      watch.reset();
-      f.convert(password);
-      return watch.elapsedMicroseconds;
-    });
-    int best = samples.fold(samples.first, min);
-    if (verbose) {
-      int delta = target - best;
-      print("[Argon2Security] t=$passes,p=$lanes,m=$memory ~ $delta us");
-    }
-    if ((strictness * target / best).round() < strictness) {
-      if (pow > 12) {
-        pow -= 2;
-      } else {
-        pow--;
-      }
-      break;
-    }
-  }
-
-  // found the maximum memory
-  memory = 1 << pow;
-
-  // now maximize the passes
-  for (passes++;; passes++) {
-    var samples = List.generate(10, (_) {
-      var f = Argon2(
-        salt: salt,
-        hashLength: hashLength,
-        type: type,
-        version: version,
-        iterations: passes,
-        parallelism: lanes,
-        memorySizeKB: memory,
-      );
-      watch.reset();
-      f.convert(password);
-      return watch.elapsedMicroseconds;
-    });
-    int best = samples.fold(samples.first, min);
-    if (verbose) {
-      int delta = target - best;
-      print("[Argon2Security] t=$passes,p=$lanes,m=$memory ~ $delta us");
-    }
-    if ((strictness * target / best).round() < strictness) {
-      passes--;
-      break;
-    }
-  }
-  if (passes == 1 && pow > 10) {
-    pow++;
-    memory = 1 << pow;
-  }
-
-  if (verbose) {
-    print('[Argon2Security] ------------');
-    print('[Argon2Security] t: $passes');
-    print('[Argon2Security] p: $lanes');
-    print('[Argon2Security] m: $memory (2^$pow)');
-  }
-
-  return Argon2Security('optimized', m: memory, t: passes, p: lanes);
 }