+++ lib/src/algorithms/keccak/keccak_64bit.dart
@@ -74,7 +74,7 @@ const List<int> _rc = <int>[
 class KeccakHash extends BlockHashSink {
   final int stateSize;
   final int paddingByte;
-  late final Uint64List qstate;
+  late final Uint64List qstate = Uint64List.view(buffer.buffer);
 
   @override
   final int hashLength;
@@ -83,16 +83,12 @@ class KeccakHash extends BlockHashSink {
     required this.stateSize,
     required this.paddingByte,
     int? outputSize, // equals to state size if not provided
-  })  : hashLength = outputSize ?? stateSize,
+  })  : assert(stateSize >= 0 && stateSize < 100),
+        hashLength = outputSize ?? stateSize,
         super(
           200 - (stateSize << 1), // rate as blockLength
           bufferLength: 200, // 1600-bit state as buffer
-        ) {
-    if (stateSize < 0 || stateSize >= 100) {
-      throw ArgumentError('The state size is not valid');
-    }
-    qstate = Uint64List.view(buffer.buffer);
-  }
+        );
 
   @override
   void reset() {